requirements.txt
# Web framewroks
fastapi
uvicorn[standard]
python-multipart 

# langchain LLM
langchain
langchain-community
langchain-core
langchain-ollama

# vectorstore
chromadb   

# Embeddings
sentence-transformers   # all-MiniLM-L6-v2

# PDF parsing
pypdf

# Environment Variables : 
python-dotenv

# Typing 
pydantic
requests
tqdm

# Authentication
pymongo
bcrypt